+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Models/ApplicationLogic.cs
@@ -2,22 +2,26 @@ using Newtonsoft.Json;
 
 namespace Altinn.Platform.Storage.Interface.Models
 {
+    /// <summary>
+    /// Represents an object with information about how the data type is handled by the application logic.
+    /// </summary>
     public class ApplicationLogic
     {
         /// <summary>
-        /// If true the app-backend will attemt to automatically create (or prefill) this data element when the task referred by taskId starts.
+        /// Gets or sets a value indicating whether the app-backend will attempt to automatically create (or prefill)
+        /// this data type when the task referred by taskId starts.
         /// </summary>
         [JsonProperty(PropertyName = "autoCreate")]
         public bool? AutoCreate;
 
         /// <summary>
-        /// Reference to the class that is started to instantiate the data type.
+        /// Gets or sets the class type to instantiate when creating an instance of this data type.
         /// </summary>
         [JsonProperty(PropertyName = "classRef")]
         public string ClassRef;
 
         /// <summary>
-        /// Reference to the schema that defines the data type.
+        /// Gets or sets the name and path to the data type schema.
         /// </summary>
         [JsonProperty(PropertyName = "schemaRef")]
         public string SchemaRef;
