+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Models/InstanceEvent.cs
@@ -1,52 +1,54 @@
 using System;
+using System.Collections.Generic;
+
 using Newtonsoft.Json;
 
 namespace Altinn.Platform.Storage.Interface.Models
 {
     /// <summary>
-    /// Model for the instance event.
+    ///Represents the an event related to an instance.
     /// </summary>
     [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class InstanceEvent
     {
         /// <summary>
-        /// unique identifier of the event.
+        /// Unique identifier of the event.
         /// </summary>
         [JsonProperty(PropertyName = "id")]
         public Guid? Id { get; set; }
 
         /// <summary>
-        /// the instance the event refers to {instanceOwnerId}/{instanceGuid}.
+        /// The instance the event is referring to {instanceOwnerId}/{instanceGuid}.
         /// </summary>
         [JsonProperty(PropertyName = "instanceId")]
         public string InstanceId { get; set; }
 
         /// <summary>
-        /// the data element which the event refers to, or null.
+        /// The data element which the event refers to, or null.
         /// </summary>
         [JsonProperty(PropertyName = "dataId")]
         public string DataId { get; set; }
 
         /// <summary>
-        /// event creation date-time
+        /// Event creation date-time
         /// </summary>
         [JsonProperty(PropertyName = "created")]
         public DateTime? Created { get; set; }
 
         /// <summary>
-        /// the event type, e.g. created, saved, process-change.
+        /// The event type, e.g. created, saved, process-change.
         /// </summary>
         [JsonProperty(PropertyName = "eventType")]
         public string EventType { get; set; }
 
         /// <summary>
-        /// the instance owner party id.
+        /// The instance owner party id.
         /// </summary>
         [JsonProperty(PropertyName = "instanceOwnerPartyId")]
         public string InstanceOwnerPartyId { get; set; }
 
         /// <summary>
-        /// the user who triggered the event.
+        /// The user who triggered the event.
         /// </summary>
         [JsonProperty(PropertyName = "user")]
         public PlatformUser User;
@@ -64,5 +66,22 @@ namespace Altinn.Platform.Storage.Interface.Models
         {
             return JsonConvert.SerializeObject(this);
         }
+    }
+
+    /// <summary>
+    /// Represents a container object with a list of events related to instances.
+    /// </summary>
+    /// <remarks>
+    /// This should be used only when an API endpoint would otherwise return a list of instance events.
+    /// Not when the list is a property of a separate class.
+    /// </remarks>
+    [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
+    public class InstanceEvents
+    {
+        /// <summary>
+        /// The actual list of instance events.
+        /// </summary>
+        [JsonProperty(PropertyName = "data")]
+        public List<InstanceEvent> Data { get; set; }
     }
 }
