+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Models/InstanceOwner.cs
@@ -4,29 +4,28 @@ using Newtonsoft.Json;
 namespace Altinn.Platform.Storage.Interface.Models
 {
     /// <summary>
-    /// Model to hold an instance owner element.
+    /// Represents information to identify the owner of an instance.
     /// </summary>
     [Serializable]
     [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class InstanceOwner
     {
         /// <summary>
-        /// The party id of the instance owner (also called instance owner party id). 
+        /// Gets or sets the party id of the instance owner (also called instance owner party id). 
         /// </summary>
         [JsonProperty(PropertyName = "partyId")]
         public string PartyId { get; set; }
 
         /// <summary>
-        /// Person number (national identification number) of the party. Null if the party is not a person.
+        /// Gets or sets person number (national identification number) of the party. Null if the party is not a person.
         /// </summary>
         [JsonProperty(PropertyName = "personNumber")]
         public string PersonNumber { get; set; }
 
         /// <summary>
-        /// Organisation number of the party. Null if the party is not an organisation.
+        /// Gets or sets the organisation number of the party. Null if the party is not an organisation.
         /// </summary>
         [JsonProperty(PropertyName = "organisationNumber")]
         public string OrganisationNumber { get; set; }
-
     }
 }
