+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Models/InstanceStatus.cs
@@ -1,32 +1,31 @@
-using Newtonsoft.Json;
 using System;
 
+using Newtonsoft.Json;
 
 namespace Altinn.Platform.Storage.Interface.Models
 {
     /// <summary>
-    /// Holds the state of an instance with respect to the Altinn inbox.
+    /// Represents set of possible statuses that dictate where the message box will display the instance.
     /// </summary>
     [JsonObject(ItemNullValueHandling = NullValueHandling.Ignore)]
     public class InstanceStatus
     {
         /// <summary>
-        /// The date the instance was archived.
+        /// Gets or sets the date the instance was archived.
         /// </summary>
         [JsonProperty(PropertyName = "archived")]
         public DateTime? Archived { get; set; }
 
         /// <summary>
-        /// The date the instance was deleted. 
+        /// Gets or sets the date the instance was deleted. 
         /// </summary>
         [JsonProperty(PropertyName = "softDeleted")]
         public DateTime? SoftDeleted { get; set; }
 
         /// <summary>
-        /// The data the instance was marked for hard delete by user.
+        /// Gets or sets the date the instance was marked for hard delete by user.
         /// </summary>
         [JsonProperty(PropertyName = "hardDeleted")]
         public DateTime? HardDeleted { get; set; }
-
     }
 }
