+++ src/Altinn.Platform/Altinn.Platform.Storage/Storage.Interface/Models/PlatformUser.cs
@@ -9,25 +9,25 @@ namespace Altinn.Platform.Storage.Interface.Models
     public class PlatformUser
     {
         /// <summary>
-        /// The identifier of the user as given by Altinn, or null if the user is a organisation user.
+        /// Gets or sets the unique user id as given by Altinn, or null if the user is a organisation user.
         /// </summary>
         [JsonProperty(PropertyName = "userId")]
         public int? UserId { get; set; }
 
         /// <summary>
-        /// The altinn org identifier of the organisation that were identified by maskinporten. 
+        /// Gets or sets the altinn org identifier of the organisation that were identified by maskinporten. 
         /// </summary>
         [JsonProperty(PropertyName = "orgId")]
         public string OrgId { get; set; }
 
         /// <summary>
-        /// the authentication level for the user which triggered the event
+        /// Gets or sets the authentication level for the user which triggered the event
         /// </summary>
         [JsonProperty(PropertyName = "authenticationLevel")]
         public int AuthenticationLevel { get; set; }
 
         /// <summary>
-        /// the end user system that were used triggered the event.
+        /// Gets or sets the end user system that were used triggered the event.
         /// </summary>
         [JsonProperty(PropertyName = "endUserSystemId")]
         public int? EndUserSystemId { get; set; }
